refactor(desktop): move resolveActivePaneView under registry/utils/
AGENTS.md groups shared utility functions under utils/; relocate so
the registry matches the convention.
